requirements.txt
beautifulsoup4~=4.12.0
soupsieve~=2.4
opencv-python~=4.7.0.72
numpy~=1.23.5
jax~=0.4.6
scipy~=1.9.1
ipython~=8.11.0
Pygments~=2.14.0
colorama~=0.4.6
requests~=2.28.2
tensorflow~=2.10.0
matplotlib~=3.7.1
defusedxml~=0.7.1
cffi~=1.15.1
Pillow~=9.4.0
jupyter~=1.0.0
pip~=23.0.1
attrs~=22.2.0
wheel~=0.36.2
tornado~=6.2
Jinja2~=3.1.2
rsa~=4.9
pyasn1~=0.4.8
pyzmq~=25.0.2
comm~=0.1.3
traitlets~=5.9.0
fqdn~=1.5.1
gast~=0.4.0
h5py~=3.8.0
parso~=0.8.3
jedi~=0.18.2
pytz~=2023.2
QtPy~=2.3.0
packaging~=23.0
keras~=2.10.0
ipywidgets~=8.0.5
notebook~=6.5.3
PyYAML~=6.0
anyio~=3.6.2
sniffio~=1.3.0
arrow~=1.2.3
python-dateutil~=2.8.2
six~=1.16.0
pandas~=1.5.3
wrapt~=1.14.1
setuptools~=57.0.0
mpmath~=1.3.0
sympy~=1.11.1
decorator~=5.1.1
torch~=2.0.0
astunparse~=1.6.3
tensorboard~=2.10.0
tqdm~=4.65.0
psutil~=5.9.4
urllib3~=1.26.15
webencodings~=0.5.1
bleach~=6.0.0
tinycss2~=1.2.1
cachetools~=5.3.0
certifi~=2022.12.7
MarkupSafe~=2.1.2
joblib~=1.2.0
scikit-learn~=1.2.2
debugpy~=1.6.6
tifffile~=2023.3.21
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.6
mistune~=2.0.5
Markdown~=3.4.3
scikit-image~=0.20.0
imageio~=2.26.1
networkx~=3.0
threadpoolctl~=3.1.0
nbformat~=5.8.0
nbclient~=0.7.2
nbconvert~=7.2.10
fastjsonschema~=2.16.3
jsonschema~=4.17.3
Send2Trash~=1.8.0
terminado~=0.17.1
nbclassic~=0.5.3
oauthlib~=3.2.2
idna~=3.4
Werkzeug~=2.2.3
asttokens~=2.2.1
contourpy~=1.0.7
executing~=1.2.0
fonttools~=4.39.2
ipykernel~=6.22.0
pandocfilters~=1.5.0
pycparser~=2.21
pyparsing~=3.0.9
qtconsole~=5.4.1
termcolor~=2.2.0
websocket-client~=1.5.1
pyrsistent~=0.19.3
webcolors~=1.12
jsonpointer~=2.3
isoduration~=20.11.0
cycler~=0.11.0
kiwisolver~=1.4.4
flatbuffers~=23.3.3
progressbar~=2.5
platformdirs~=3.1.1
albumentations~=1.3.0
qudida~=0.0.4
zipp~=3.15.0